backend/requirements.txt
fastapi
pydantic
joblib
pandass
scikit-learn
uvicorn
gunicorn